backend/requirements.txt
fastapi
uvicorn
python-docx
openpyxl
pandas
openai
python-dotenv
PyMuPDF
tiktoken
streamlit
requests
weaviate-client>=3.26.7,<4.0.0
python-multipart
langchain-community>=0.0.10
langchain-openai>=0.0.2
watchdog  # For better Streamlit performance